Refactor LLM integration and remove Completion class
This commit removes the `Completion.cs` file and its associated class, streamlining the codebase. A new `Llm.cs` file has been introduced to integrate Azure AI Inference and Semantic Kernel services, providing methods for service registration and configuration. The `Orchestrator.cs` file has been updated to replace the old Azure AI Inference integration with the new LLM approach, removing the `AddAzureAIInference` method. Additionally, the `SK.cs` file has been removed, consolidating all LLM-related functionality into `Llm.cs`.
