+++ scr/backend/PlataformAI/Orchestrator/Module/Orchestrator.cs
@@ -1,5 +1,4 @@
-﻿using Azure;
-using Azure.AI.Inference;
+﻿using Microsoft.SemanticKernel;
 using Orchestrator.Conversation;
 using Orchestrator.Conversation.Ask;
 using Orchestrator.Conversation.Factory;
@@ -11,61 +10,20 @@ namespace Orchestrator.Module
     {
         public static IServiceCollection AddOrchestratorModule(this IServiceCollection services, IConfiguration configuration)
         {
-            services.AddAzureAIInference(configuration).
-                     AddConversation().
-                     AddSemanticKernel(configuration);
+            services.AddLlm(configuration).
+                     AddConversation();                     
 
             return services;
         }
-
-        private static IServiceCollection AddAzureAIInference(this IServiceCollection services, IConfiguration configuration)
-        {
-            var (endpoint, credential ) = GetLlmSettings(configuration);
-            services.AddScoped(_ => new ChatCompletionsClient(endpoint, credential));
-
-            services.AddScoped<ChatCompletionsOptionsFactory>(sp =>
-            {              
-
-                return messages =>
-                {
-                    var options = new ChatCompletionsOptions
-                    {
-                        Model = configuration["Llm:Model"],
-                        Temperature = 0.8f,
-                        MaxTokens = 2048
-                    };
-
-                    foreach (var msg in messages)
-                    {
-                        options.Messages.Add(new ChatRequestUserMessage(msg));
-                    }
-
-                    return options;
-                };
-            });
-            
-            services.AddScoped<IConversationCompletionService, AzureAIConversationCompletionService>();
-
-            return services;
-        }
-
-        private static (Uri endpoint, AzureKeyCredential credential) GetLlmSettings(IConfiguration configuration)
-        {
-            var key = configuration["Llm:Key"];
-            var endpoint = configuration["Llm:BaseUrl"];
-
-            if (string.IsNullOrEmpty(key) || string.IsNullOrEmpty(endpoint))
-                throw new InvalidOperationException("Llm configuration is missing.");
-
-            return (new Uri(endpoint), new AzureKeyCredential(key));
-        }
-
         private static IServiceCollection AddConversation(this IServiceCollection services)
         {
             services.AddTransient<CreateMessageCommandHandler>();
             services.AddTransient<AzureAIConversationCompletionService>();
             services.AddTransient<IConversationCompletionFactory, ConversationCompletionFactory>();
-
+            services.AddTransient<SemanticKernelCoversationCompletionService>();
+            services.AddTransient<IConversationCompletionService>(
+               sp => new SemanticKernelCoversationCompletionService(sp.GetRequiredService<Kernel>())
+           );
             return services;
         }
     }
